[ADD] stock_picking_transfer_lot_autoassign (#289)
===================================
Auto-assignment of lots on pickings
===================================

When working with lots, it's very uncomfortable to introduce the quantity,
lot by lot, when transferring pickings from your warehouse (outgoing or
internal).

This module automatically assigns the reserved quantity as the done one, so
that you only have to change it in case of divergence, but having the
possibility of transferring directly.

Configuration
=============

* Make sure you have selected the proper removal strategy on your product
  categories.
* Configure the product on the page "Inventory", field "Tracking" with one of
  these values: "By Unique Serial Number" or "By Lots".

Usage
=====

* Create an outgoing or an internal picking.
* Include one product with lots and with enough stock.
* Click on "Mark as Todo" button, and then on "Reserve".
* Clicking on the icon with the three items bullet list on the "Operations"
  tab you will see that the quantities have been auto-assigned on the "Done"
  column.